arreglo de test y mapper y otros arreglos

+++ src/main/java/com/proyecto/reservaVuelos/dto/VueloModelDto.java
@@ -25,7 +25,7 @@ public class VueloModelDto {
 
     private double precio;
 
-    private long asientos;
+    private int asientos;
 
     private String tipoVuelo;
